pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-recap"
version = "0.1.0"
description = "Transcribe videos and extract visual context from recordings using OpenAI Whisper"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Ari Mahpour" },
]
keywords = ["whisper", "transcription", "video", "ffmpeg", "subtitles"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "openai-whisper",
]

[project.scripts]
open-recap = "open_recap.cli:main"

[project.urls]
Homepage = "https://github.com/amahpour/open-recap"
Repository = "https://github.com/amahpour/open-recap"
Issues = "https://github.com/amahpour/open-recap/issues"

[tool.setuptools.packages.find]
where = ["src"]
